app/requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.10.3
